requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.9.2
sqlalchemy==2.0.35
databases==0.9.0
asyncmy==0.2.9
pymysql==1.1.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
pandas==2.2.3
openpyxl==3.1.5
python-dotenv==1.0.1